optimize get collection items
