+++ packages/lightnet/src/content/query-media-items.ts
@@ -1,3 +1,4 @@
+import { compareMediaCollectionItems } from "./compare-media-collection-items"
 import type { MediaItemEntry } from "./content-schema-internal"
 
 export type MediaItemQuery<TMediaItem extends MediaItemEntry> = {
@@ -59,31 +60,8 @@ export const queryMediaItems = async <TMediaItem extends MediaItemEntry>(
   }
   const { collection } = where
   if (!orderBy && collection) {
-    items.sort((a, b) => compareCollectionItems(a, b, collection))
+    items.sort((a, b) => compareMediaCollectionItems(a, b, collection))
   }
 
   return items.slice(0, limit)
-}
-
-function compareCollectionItems(
-  item1: MediaItemEntry,
-  item2: MediaItemEntry,
-  collectionId: string,
-) {
-  const getIndex = (item: MediaItemEntry) =>
-    item.data.collections?.find(
-      ({ collection }) => collection.id === collectionId,
-    )?.index
-  const index1 = getIndex(item1)
-  const index2 = getIndex(item2)
-  if (index1 === index2) {
-    return item1.id.localeCompare(item2.id)
-  }
-  if (index1 === undefined && index2 !== undefined) {
-    return 1
-  }
-  if (index1 !== undefined && index2 === undefined) {
-    return -1
-  }
-  return index1! - index2!
 }
